+++ examples/1INFILTR/solve.py
@@ -4,7 +4,6 @@
 
 from __future__ import division, absolute_import, print_function
 
-import numpy as np
 import matplotlib.pyplot as plt
 
 from fronts import solve